+++ ui/cypress/e2e/buckets.test.ts
@@ -31,6 +31,20 @@ describe('Buckets', () => {
       })
 
       cy.getByTestID(`bucket--card--name ${newBucket}`).should('exist')
+
+      // Add a label
+      cy.getByTestID('inline-labels--add')
+        .first()
+        .click()
+
+      const labelName = 'l1'
+      cy.getByTestID('inline-labels--popover--contents').type(labelName)
+      cy.getByTestID('inline-labels--create-new').click()
+      cy.getByTestID('create-label-form--submit').click()
+
+      // Delete the label
+      cy.getByTestID(`label--pill--delete ${labelName}`).click({force: true})
+      cy.getByTestID('inline-labels--empty').should('exist')
     })
 
     it("can update a bucket's retention rules", () => {
@@ -60,28 +74,28 @@ describe('Buckets', () => {
 
     describe('Searching and Sorting', () => {
       it('can sort by name and retention', () => {
-        cy.getByTestID('name-sorter').click()
-        cy.getByTestID('bucket-card')
-          .first()
-          .contains('defbuck')
+        const buckets = ['defbuck', '_tasks', '_monitoring']
+        cy.getByTestID('name-sorter')
+          .click()
+          .then(() => {
+            cy.get('[data-testid*="bucket-card"]').each((val, index) => {
+              const testID = val.attr('data-testid')
+              expect(testID).to.include(buckets[index])
+            })
+          })
 
-        cy.getByTestID('name-sorter').click()
-        cy.getByTestID('bucket-card')
-          .first()
-          .contains('_monitoring')
-
-        cy.getByTestID('retention-sorter').click()
-        cy.getByTestID('bucket-card')
-          .first()
-          .contains('_tasks')
-
-        cy.getByTestID('retention-sorter').click()
-        cy.getByTestID('bucket-card')
-          .first()
-          .contains('defbuck')
+        cy.getByTestID('name-sorter')
+          .click()
+          .then(() => {
+            const asc_buckets = buckets.slice().sort()
+            cy.get('[data-testid*="bucket-card"]').each((val, index) => {
+              const testID = val.attr('data-testid')
+              expect(testID).to.include(asc_buckets[index])
+            })
+          })
 
         cy.getByTestID('search-widget').type('tasks')
-        cy.getByTestID('bucket-card').should('have.length', 1)
+        cy.get('.cf-resource-card').should('have.length', 1)
       })
     })